// src/app/api/ui/models/participant-offer.interface.ts
import { Category } from './category.interface';
import { CharacteristicValues } from './characteristic-values.interface';
import { Country } from './country.interface';
import { DeliveryRegionInfo } from './delivery-region-info.interface';
import { ManufacturerCountryRegistryTypeEnum } from '../enums/manufacturer-country-registry-type.enum';
import { OfferCreatedMethodEnum } from '../enums/offer-created-method.enum';
import { Okpd2Market } from './okpd-2-market.interface';
import { ParticipantOfferDocument } from './participant-offer-document.interface';
import { ParticipantOfferSourceEnum } from '../enums/participant-offer-source.enum';
import { ParticipantOfferSpecificFlagEnum } from '../enums/participant-offer-specific-flag.enum';
import { ParticipantOfferStateEnum } from '../enums/participant-offer-state.enum';
import { ParticipantOfferTypeEnum } from '../enums/participant-offer-type.enum';
import { PriceRegionDetail } from './price-region-detail.interface';
import { RegionKladr } from './region-kladr.interface';
import { TruTypeEnum2 } from '../enums/tru-type-enum-2.enum';
import { UsedClassificatorTypeEnum } from '../enums/used-classificator-type.enum';
import { VatRateStateEnum } from '../enums/vat-rate-state.enum';

// @ts-ignore
export interface ParticipantOffer {
  category: Category;
  characteristicFileGuid: string;
  characteristicFromUi: boolean;
  characteristicValues: CharacteristicValues[];
  classificatorCode: string;
  classificatorDescription: string;
  country: Country;
  createdMethod: OfferCreatedMethodEnum;
  currencyCode: string;
  deliveryCost: number;
  deliveryPlace: string;
  deliveryRegionInfos: DeliveryRegionInfo[];
  deliveryRegions: string;
  duringDays: number;
  exclusiveToTrade: boolean;
  externalSystemIdentifier: string;
  externalSystemIdentifierNumeric: number;
  id: number;
  isPublicOffer: boolean;
  itemInStock: boolean;
  ksrCode: string;
  lotItemId: number;
  manufacturer: string;
  manufacturerCountryRegistryNumber: string;
  manufacturerCountryRegistryType: ManufacturerCountryRegistryTypeEnum;
  manufacturerCountryScore: number;
  minQuantity: number;
  neverExpired: boolean;
  offerState: ParticipantOfferStateEnum;
  offerTerms: string;
  offerType: ParticipantOfferTypeEnum;
  okeiCode: string;
  okeiDescription: string;
  okpd2Codes: Okpd2Market[];
  participantOfferDocuments: ParticipantOfferDocument[];
  participantOfferSource: ParticipantOfferSourceEnum;
  participantOfferSpecific: ParticipantOfferSpecificFlagEnum;
  possibleDeliveryBySeller: boolean;
  possibleSelfDelivery: boolean;
  possibleSendByTransportCompany: boolean;
  priceListId: number;
  priceRegionDetails: PriceRegionDetail[];
  printFormGuid: string;
  productDescription: string;
  productName: string;
  publishDateStart: string;
  quantity: number;
  recipientOrganizationInn: string;
  recipientOrganizationKpp: string;
  recipientOrganizationName: string;
  regionalProductSignChecked: boolean;
  regions: RegionKladr[];
  requestDateEnd: string;
  requestDateStart: string;
  selfDeliveryPlace: string;
  signature: string;
  tradeId: number;
  truType: TruTypeEnum2;
  unitPrice: number;
  unitPriceInCurrency: number;
  usedClassificatorType: UsedClassificatorTypeEnum;
  vatRate: number;
  vatRateState: VatRateStateEnum;
}